env2/requirements.txt
pillow
opencv-python
matplotlib

